feat: :heavy_plus_sign: Add tools to top page
Issue: #93
