+++ src/app/lib.ts
@@ -1,0 +1,7 @@
+export function formatDate(createdAt: string): string {
+  const date = new Date(createdAt);
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  return `${year}年${month}月${day}日`;
+}
