+++ src/app/tools/data.json
@@ -1,0 +1,8 @@
+[
+  {
+    "slug": "hello",
+    "name": "やっはろー",
+    "description": "ハローワールドを表示するだけのページです。",
+    "createdAt": "2024-04-19"
+  }
+]
